requirements.txt
streamlit>=1.36
mediapipe>=0.10.14
opencv-python>=4.9
numpy>=1.26,<2.0
scipy>=1.11
pandas>=2.1
pyyaml>=6.0
anthropic>=0.34
python-dotenv>=1.0
dtw-python>=1.3
pillow>=10.2
boto3>=1.34
botocore>=1.34
